moved views to feature based folders
