+++ YUP/YUP.App/MainWindow.xaml
@@ -1,10 +1,13 @@
-﻿<Controls:MetroWindow x:Class="YUP.App.MainWindow"
+﻿<Controls:MetroWindow 
+        x:Class="YUP.App.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:YUP.App"
-        xmlns:views="clr-namespace:YUP.App.Views"
+        xmlns:videos="clr-namespace:YUP.App.Videos"
+        xmlns:sidecontrols="clr-namespace:YUP.App.SideControls"
+        xmlns:player="clr-namespace:YUP.App.Player"
         mc:Ignorable="d"
         Title="YUP - Youtube Universal Player"
         Width="1020"
@@ -39,20 +42,20 @@
         <Grid x:Name="GridControls" 
               Grid.Column="2"
               Grid.Row="0">
-            <views:SideControlsView></views:SideControlsView>
+            <sidecontrols:SideControlsView></sidecontrols:SideControlsView>
         </Grid>
                
               
         <Grid x:Name="GridDetails"
               Grid.Column="1"
               Grid.Row="0">
-            <views:VideosView></views:VideosView>
+            <videos:VideosView></videos:VideosView>
         </Grid>
 
         <Grid x:Name="GridPlayer"
               Grid.Column="0"
               Grid.Row="0">
-            <views:Player></views:Player>
+            <player:Player></player:Player>
         </Grid>
 
     </Grid>
