+++ YUP/YUP.App/Player/PlayerView.xaml
@@ -1,9 +1,8 @@
-﻿<UserControl x:Class="YUP.App.Views.Player"
+﻿<UserControl x:Class="YUP.App.Player.Player"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:YUP.App.Views"
              mc:Ignorable="d" 
              d:DesignHeight="300" d:DesignWidth="300">
     <Grid>
