+++ YUP/YUP.App/SideControls/SideControlsView.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System.Windows.Controls;
+
+namespace YUP.App.SideControls
+{
+    /// <summary>
+    /// Interaction logic for SideControlsView.xaml
+    /// </summary>
+    public partial class SideControlsView : UserControl
+    {
+        public SideControlsView()
+        {
+            InitializeComponent();
+        }
+    }
+}
