+++ YUP/YUP.App/Videos/VideosView.xaml
@@ -1,9 +1,8 @@
-﻿<UserControl x:Class="YUP.App.Views.VideosView"
+﻿<UserControl x:Class="YUP.App.Videos.VideosView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:YUP.App.Views"
              xmlns:controls="http://metro.mahapps.com/winfx/xaml/controls"
              mc:Ignorable="d" 
              d:DesignHeight="300" d:DesignWidth="300">
