+++ YUP/YUP.App/Videos/VideosView.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System.Windows.Controls;
+
+namespace YUP.App.Videos
+{
+    /// <summary>
+    /// Interaction logic for VideosView.xaml
+    /// </summary>
+    public partial class VideosView : UserControl
+    {
+        public VideosView()
+        {
+            InitializeComponent();
+        }
+    }
+}
