+++ YUP/YUP.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 14.0.24720.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Player", "Player", "{3BCC25B0-B670-49AE-9C4B-C12806440491}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "YUP.Player.Youtube", "YUP.Player.Youtube\YUP.Player.Youtube.csproj", "{24B37DAB-2A4C-4306-9363-0B5CAD7D4DCA}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "App", "App", "{427AFE3B-F13C-45E6-904F-DA703D4D1360}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "YUP.App", "YUP.App\YUP.App.csproj", "{FF14D62A-9631-456D-9928-C2BC846D8265}"
@@ -17,10 +15,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{24B37DAB-2A4C-4306-9363-0B5CAD7D4DCA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{24B37DAB-2A4C-4306-9363-0B5CAD7D4DCA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{24B37DAB-2A4C-4306-9363-0B5CAD7D4DCA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{24B37DAB-2A4C-4306-9363-0B5CAD7D4DCA}.Release|Any CPU.Build.0 = Release|Any CPU
 		{FF14D62A-9631-456D-9928-C2BC846D8265}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{FF14D62A-9631-456D-9928-C2BC846D8265}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FF14D62A-9631-456D-9928-C2BC846D8265}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -30,7 +24,6 @@ Global
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{24B37DAB-2A4C-4306-9363-0B5CAD7D4DCA} = {3BCC25B0-B670-49AE-9C4B-C12806440491}
 		{FF14D62A-9631-456D-9928-C2BC846D8265} = {427AFE3B-F13C-45E6-904F-DA703D4D1360}
 	EndGlobalSection
 EndGlobal
